Move away from using the deprecated 'imp' package
While there is a [recipe][] in the importlib docs for importing a
file, it seems that that only works well for single files. Since
we want to allow sibling files to be used this path fudging seems
to still be the way to go.

Fixes https://github.com/PeterJCLaw/srcomp/issues/4

[recipe]: https://docs.python.org/3/library/importlib.html#importing-a-source-file-directly