Ensure that we can load several scorers independantly
Switch to using runpy to load the scorer. It's the package which
implements running files from the command line and so is designed
for pretty much this sort of thing. We still need to do some path
munging ourselves in order to preserve support for sibling files
next to score.py, but that's ok.