Fix detection of whether a commit is a parent of another commit
Previously this worked fine for linear relationships, but failed
to account for sibling commits, where `git rev-list` will return
the list of commits that one commit has diverged from the branch
that the other one is on (even if that is only the one commit
itself). The simple fix is to call rev-list twice so we can see
both sides of the relationship. It feels like there ought to be
a better approach here, but this works for now (and is now tested).