Support static knockout rounds, matches & positions > 9
Previously the static knockout scheduler relied on the input string
being exactly three characters, which in turn forced the indices
of the rounds, matches & positions being referred to to single
digits. This was historically not an issue, however for SR2025
we're exploring a double-elimination knockout which needs more
than 10 rounds.

This adds support for that by introducing a format which does not
rely on the number of characters in the input.